+++ frb_example/pure_dart/lib/src/rust/api/optional_primitive_misc.dart
@@ -4,10 +4,6 @@ import '../frb_generated.dart';
 import 'package:flutter_rust_bridge/flutter_rust_bridge_for_generated.dart';
 
 Future<int?> primitiveOptionalTypes(
-        {int? myI32,
-        BigInt? myI64,
-        double? myF64,
-        bool? myBool,
-        dynamic hint}) =>
+        {int? myI32, int? myI64, double? myF64, bool? myBool, dynamic hint}) =>
     RustLib.instance.api.primitiveOptionalTypes(
         myI32: myI32, myI64: myI64, myF64: myF64, myBool: myBool, hint: hint);
